% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_functions.R
\name{iteratePseudoBulk}
\alias{iteratePseudoBulk}
\title{This function iterates 'pseudobulk' DE analysis using DESeq2 pipeline \
on scRNAseq data for genes of interest.
It is essentially bootstrapping for scRNAseq pseudobulk analysis.}
\usage{
iteratePseudoBulk(
  sce,
  cellIDs = NULL,
  obs,
  contrasts,
  replicates = NULL,
  alpha = 0.05,
  genes = NULL,
  iterations = 10
)
}
\arguments{
\item{sce:}{SCE object, takes raw counts}

\item{cellIDs:}{IDs of cells of interest belonging to cluster(s)}

\item{obs:}{obs/metadata variable to group by - must column name in be in sce colData}

\item{replicates:}{number of artificial replicates to split each condition into, recommend at least 500 cells per replicate after splitting}

\item{contrast:}{DESeq contrast for results - take a vector of 2, eg. c('treatment','control')}

\item{alpha:}{significance level}

\item{genes:}{vector of genes of interest, if NULL, iterate and test for all}

\item{iterations:}{number of times to permute random sampling and results function}
}
\value{
A list of LogFoldChange, adjusted P value, and p value statistics for genes of interest
}
\description{
This function iterates 'pseudobulk' DE analysis using DESeq2 pipeline \
on scRNAseq data for genes of interest.
It is essentially bootstrapping for scRNAseq pseudobulk analysis.
}
