% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_functions.R
\name{runPseudoBulk}
\alias{runPseudoBulk}
\title{This function performs pseudobulk analysis on scRNAseq data by random sampling}
\usage{
runPseudoBulk(
  sce,
  cellIDs = NULL,
  obs,
  contrasts,
  replicates = NULL,
  alpha = 0.05
)
}
\arguments{
\item{sce:}{SCE object, takes raw counts}

\item{cellIDs:}{IDs of cells of interest belonging to cluster(s)}

\item{obs:}{obs/metadata variable to group by - must column name in be in sce colData}

\item{replicates:}{number of artificial replicates to split each condition into, recommend at least 500 cells per replicate after splitting}

\item{contrasts:}{DESeq contrast for results - take a vector of 2, eg. c('treatment','control')}

\item{alpha:}{DESeq alpha parameter (significance level for DE testing)}
}
\value{
: list of Pseudobulk raw counts, DESeq2 object, VST-transformed matrix, DE results, colData design table
}
\description{
This function performs pseudobulk analysis on scRNAseq data by random sampling
}
